KAFKA-18265: Move inflight batch and state classes from SharePartition (2/N)
